transaksi-setrika

+++ app/Repositories/BaseRepository.php
@@ -49,28 +49,34 @@ class BaseRepository implements BaseRepositoryInterface {
         return $this->model->create($attributes);
     }
 
-    public function update($id, $attributes, $isFile = false, $field = null, $folder = null) {
-        if (is_numeric($id)) {
-            $model = $this->model->find($id);
-        } else {
-            $model = $this->model->where('key', $id)->first();
+    public function update($params, $attributes, $isFile = false, $field = null, $folder = null) {
+        if (is_numeric($params)) {
+            $model = $this->model->find($params);
+        } else if (is_array($params)) {
+            $model = $this->model->where($params)->first();
         }
-        if ($isFile == true) {
-            if (isset($field)) {
-                foreach ($field as $key => $value) {
-                    if (request()->file($value)) {
-                        File::delete('storage/' . $model[$value]);
-                        $attributes[$value] = request()->file($value)->store('filemanager/' . $folder, 'public');
+
+        if ($model) {
+            if ($isFile == true) {
+                if (isset($field)) {
+                    foreach ($field as $key => $value) {
+                        if (request()->file($value)) {
+                            File::delete('storage/' . $model[$value]);
+                            $attributes[$value] = request()->file($value)->store('filemanager/' . $folder, 'public');
+                        }
                     }
                 }
             }
-        }
-        if (is_numeric($id)) {
-            $model->update($attributes);
+            if (is_numeric($params)) {
+                $model->update($attributes);
+            } else if (is_array($params)) {
+                $model->where($params)->update($attributes);
+            }
+            return $model;
         } else {
-            $model->where('key', $id)->update($attributes);
+            return false;
         }
-        return $model;
+
     }
 
     public function softDelete($id, $foreign_key = null) {
